requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.1.1
djangorestframework==3.14.0
drf-jwt==1.19.2
gunicorn==20.1.0
pycparser==2.21
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
types-cryptography==3.3.23
